// src/api/response.type.ts
export interface BaseListType {
  label: string
  value: string | number
}
// ...
export interface UserInfo {
  username: string
  real_name: string
  nickname: string
  email: string
  phone: string
  gender: GENDER.Enums
  date_of_birth: string
  invite_code: string
  contact: {
    contact1: string
    contact2: string
  }
  empty_password?: boolean
  member_level?: number
  has_withdrawal_password?: boolean
  avatar_path: string
  self_exclusion_at: number | null
  approval_status: boolean
  login_provider?: LOGIN_PROVIDER.Enums // 會員註冊的方式
  login_type?: LOGIN_PROVIDER.Enums // 當前會員登入的方式
  is_member_agent: boolean
}

export interface GetUserInfo {
  account: string // 帳號
  ref_account: string // 推薦人
  invite_code: string // 推薦人
  fullname: string // 真實姓名
  nickname: string // 暱稱
  dob: string // 生日
  gender: GENDER.Enums // 性別
  country: string // 國碼
  phone: string // 電話
  email: string // 電子郵箱
  sns_account_1: string // 通訊軟體1
  sns_account_2: string // 通訊軟體2
  nationality: string // 國籍
  place_of_birth: string // 出生地
  present_address: string // 現居地址
  permanent_address: string // 永久地址
  nature_of_work: string // 工作性質
  source_of_income: string // 收入來源
  gaming_site: number // 遊戲場地
  empty_password: boolean
  member_leve: number
  has_withdrawal_password: boolean
  avatar_path: string
  self_exclusion_at: number | null
  member_level: number
  approval_status: boolean
  login_provider?: LOGIN_PROVIDER.Enums // 會員註冊的方式
  login_type?: LOGIN_PROVIDER.Enums // 當前會員登入的方式
  middle_name: string
  last_name: string
  city: string
  country_name: string
  region: string
  province: string
  area: string
  postal_code: string
  job_type_details: string
  income_source_details: string
  sms_otp: string
  is_member_agent: boolean
  is_sub_ad: boolean
  registered_ip: string // 註冊IP
}

export interface AccountInfo {
  [key: string]: any
  self_exclusion_at: number | null
}

export interface UserWallet {
  currency_id: number
  currency_code: string
  remaining_turnover: string
  balance: string
  in_use: boolean
  label?: string
  wallet_type: WALLET_TYPE.Enums
}

export type UserWalletList = UserWallet[]

export interface UserLevel {
  id: number
  quota: number
  img: string
  lang: string
  updated_time: string
  remark: string
}

export type UserLevelsInfo = UserLevel[]

export interface UserRemark {
  remark: string
}

export type BaseMemberColumn = Record<keyof Request.UserInfoForm, string>

export interface MemberColumn {
  column_name: string
  customize: boolean
  lang?: {
    [key: string]: string
  }
  required: boolean
  edit: boolean
  type: INPUT_TYPE.Enums
  values: BaseListType[]
  column_label?: string
}

export type MemberColumnList = MemberColumn[]
// ...
export interface GetMemberSummary {
  bet_count: number
  bet_amount: string
  bonus: string
  deposit: string
  login_at: string
  member_id: number
  prize: string
  profit: string
  valid_bet: string
  withdraw: string
  metrics: {
    date: string
    deposit: string
    withdraw: string
  }[]
}
// ...
export interface GetKYCStatus {
  status: KYC_VERIFIED.Enums
}

export interface UserKycItem {
  id: number
  img: string
  type: KYC_TYPE.Enums
  status: KYC_STATUS_CODE.Enums
  correspondence: number
  updated_time: number
  updated_at: number
  fileName: string
}

export type UserKycList = UserKycItem[]
// ...
export type GetMemberAgentQuotaList = {
  member_id: number
  member_account: string
  hierarchy_level: number
  balance: string
  remain_quota_amount: string
  is_member_agent: boolean
  is_direct_downline: boolean
  is_downline_search_available: boolean
  register_date: number
  last_login_date: number
}

export interface GetMemberAgentQuotaListData {
  list: GetMemberAgentQuotaList[]
  offset: number
  size: number
  total: number
}

export type GetLowerLevelMemberAgentQuotaList = {
  member_id: number
  member_account: string
  hierarchy_level: number
  balance: string
  remain_quota_amount: string
  is_member_agent: boolean
  is_direct_downline: boolean
  is_downline_search_available: boolean
  register_date: number
  last_login_date: number
}

export type GetLowerLevelMemberAgentQuotaListData = {
  list: GetLowerLevelMemberAgentQuotaList[]
  offset: number
  size: number
  total: number
}

export type GetMemberAgentQuotaMoneyHistory = {
  member_account: string
  id: number
  transaction_code: string
  wager_code: string
  product_name: string
  game_name: string
  promotion_title: null
  updated_at: string
  updated_at_unix: string
  currency_code: string
  action_type: number
  wallet_type: number
  amount: string
  before_balance: string
  after_balance: string
  metadata: object
  crypto_id: number
  crypto: string
  crypto_rate: string
  crypto_amount: string
}

export interface GetMemberAgentQuotaMoneyHistoryList {
  list: GetMemberAgentQuotaMoneyHistory[]
  offset: number
  size: number
  total: number
}

export type GetMemberAgentQuotaAmount = {
  currency_id: number
  remain_quota_amount: string
  total_quota_amount: string
}

export type UserBetHistoryTotalItem = {
  currency_code: string
  win_loss_amount: string
}

export type GetUserBetHistoryTotal = {
  login_at: number
  bet_historys: UserBetHistoryTotalItem[]
}

export type ProductAvailableCurrency = {
  currency_id: number
  code: string
}[]
export interface MemberCustomizeColumnList {
  code: number
  msg: string
  data: MemberCustomizeColumnListData[]
}

export interface MemberCustomizeColumnListData {
  column_name: string
  customize: boolean
  edit: boolean
  lang: {
    [key: string]: string
  }
  required: boolean
  type: number
  values: {
    label: string
    value: string
  }[]
  column_label?: string
}

export interface MemberAgentReferralList {
  code: number
  msg: string
  data: MemberAgentReferralListData[]
}

export interface MemberAgentReferralListData {
  member_id: number
  member_account: string
}

export interface MemberAgentTagListData {
  id: number
  name: string
  type: number
  remark: string
  enabled: boolean
  checked?: boolean
}

export interface MemberAgentInfo {
  [key: string]: any
}

export interface GetMemberAgentBetReportList {
  list: GetMemberAgentBetReportItem[]
  summary: GetMemberAgentBetReportSummaryItem
  offset: number
  size: number
  total: number
}

export interface GetMemberAgentBetReportItem {
  bet_amount: string
  bet_count: number
  bonus: string
  member_account: string
  member_id: number
  payout: string
  profit: string
  profit_rate: string
  valid_bet_amount: string
  win_count: number
}

export interface GetMemberAgentBetReportSummaryItem {
  page: GetMemberAgentBetReportSummaryItemPage
  total: GetMemberAgentBetReportSummaryItemTotal
}

export interface GetMemberAgentBetReportSummaryItemPage {
  bet_amount: string
  bet_count: number
  bonus: string
  payout: string
  profit: string
  profit_rate: string
  valid_bet_amount: string
  win_count: number
}

export interface GetMemberAgentBetReportSummaryItemTotal {
  bet_amount: string
  bet_count: number
  bonus: string
  payout: string
  profit: string
  profit_rate: string
  valid_bet_amount: string
  win_count: number
}

export interface GetMemberAgentWagerList {
  list: GetMemberAgentWagerListItemData[]
  summary: GetMemberAgentWagerListItemSummary
  offset: number
  size: number
  total: number
}

export interface GetMemberAgentWagerListItemData {
  bet_amount: string
  channel_code: string
  created_at: string
  game_title: string
  gaming_site: number
  gaming_site_title: string
  member_account: string
  payload: { [key: string]: any }
  payout: string
  product_code: number
  product_title: string
  profit: string
  settled_at: string
  status: number
  status_title: string
  valid_bet_amount: string
  wager_code: string
}

export interface GetMemberAgentWagerListItemSummary {
  page: GetMemberAgentWagerListItemSummaryPage
  total: GetMemberAgentWagerListItemSummaryTotal
}

export interface GetMemberAgentWagerListItemSummaryPage {
  bet_amount: string
  payout: string
  profit: string
  total: number
  valid_bet_amount: string
}

export interface GetMemberAgentWagerListItemSummaryTotal {
  bet_amount: string
  payout: string
  profit: string
  total: number
  valid_bet_amount: string
}

export interface GetMemberAgentWagerDetail {
  content: string
}

export interface GetMemberAgentReport {
  personal: {
    deposit: number
    withdraw: number
    bet_count: number
    bet_amount: number
    valid_bet: number
    prize: number
    profit: number
    rate: number
    bonus: number
  }
  team: {
    deposit: number
    withdraw: number
    bet_count: number
    bet_amount: number
    valid_bet: number
    prize: number
    profit: number
    rate: number
    bonus: number
    member_id: number
    member_account: string
    member_count: number
    currency_id: number
  }
}

export interface GetMemberTeamAgentReportList {
  deposit: number
  withdraw: number
  bet_count: number
  bet_amount: number
  valid_bet: number
  prize: number
  profit: number
  rate: number
  bonus: number
  member_id: number
  member_account: string
  member_count: number
  currency_id: number
}

export interface GetMemberTeamAgentReport {
  list: GetMemberTeamAgentReportList[]
  pagination: {
    offset: number
    size: number
    total: number
  }
}